Create Greatest Common Divisor.c
What is the greatest common divisor of 54 and 24?

The number 54 can be expressed as a product of two integers in several different ways:

1,2,3,6,9,18,27,54.\,

Similarly, the divisors of 24 are: 1,2,3,4,6,8,12,24.\,

The numbers that these two lists share in common are the common divisors of 54 and 24:
1,2,3,6.\,
The greatest of these is 6. That is, the greatest common divisor of 54 and 24. One writes:
gcd(54,24)=6.\,